bot_service/requirements.txt
python-telegram-bot==20.8
requests
Pillow